feat(domain): migrate to u128 for amount handling to support large token values

+++ crates/domain/src/models/condition.rs
@@ -19,6 +19,6 @@ pub enum Condition {
     Comparison {
         metric: Metric,
         comparator: Comparator,
-        value: u64,
+        value: u128,
     },
 }
